Retain CreateSnapshot error
The error returned from CreateSnapshot was overwritten by the call to
removeAnnVolumeSnapshotBeingCreated. Retain the error through variable
shadowing so that it can be propagated and reported upwards the call
stack.

We also improve the error line by dropping an unnecessary comma and
dequoting the error string (since we already employ a colon separator).