API change to improve VolumeGroupSnapshot restore
Change `VolumeSnapshotRefList` in `VolumeGroupSnapshotStatus` to
`PVCVolumeSnapshotRefList`, allowing users to map a `VolumeSnapshot` to
a specific `PersistentVolumeClaim`.

Change `VolumeSnapshotContentRefList` in
`VolumeGroupSnapshotContentStatus` to `PVVolumeSnapshotContentRefList`,
allowing users to map a `VolumeSnapshotContent` to a specific
`PersistentVolume`.

This two changes allow users to map each `VolumeSnapshot` back to the
original PVC it was taken from.