ensure vs-deleted annotation added to Retain content
Add bound finalizer to Retain policy snapshots to make sure the annotate logic
in checkandRemoveSnapshotFinalizersAndCheckandDeleteContent() effective.

The added finalizer does not change the current behaviour, because the source
finalizer is always added and only removed after the vs-deleted annotation is
added successfully. But we should not rely on this behaviour in case of future
changes.

The original annotate logic in syncContent() is not needed because all deleting
snapshot will go through above logic. And natually, if a content is never
bound, is will never get the vs-deleted annotation.

This also makes Retain and Delete policy snapshot more similar, which should
bring less surprises.