Avoid conflicts in updateSnapshotContentStatus
This patch replace the usage of `UpdateStatus` with a patch against the
`status` subresource in the `updateSnapshotContentStatus`.

The reason behind this change is to avoid conflicts that could
potentially arise, as they pose a risk within the context of the calling
function `createGroupSnapshotWrapper`.

If `createGroupSnapshotWrapper` is called multiple times for the same
GroupSnapshot, this will lead to multiple stale `VolumeSnapshot` and
`VolumeSnapshotContent` being created.

Related to: #1050