Correctly detect newly created Volume[Group]SnapshotClasses
The controller is using an informer to get VolumeGroupSnapshotClasses
and VolumeSnapshotClasses. Unfortunately, objects returned by that will
not have the API Kind info set.

This prevented the IsDefaultAnnotation function from working correctly.

This patch avoid relying on the object Kind to know whether a
VolumeGroupSnapshotClass or a VolumeSnapshotClass is set as default.

See: https://github.com/kubernetes/client-go/issues/308

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>