sidecar: Get credentials from annotation for snapshots
When a snapshotcontent is created as a result of a volumegroupsnapshot
the status for it is updated by calling the ListSnapshots RPC. Some COs
require the credentials to be present in order to serve this RPC.

The current code tries to get credentials only from a volumesnapshotclass
if present. This condition will never be met in case of VGS.

This patch adapts to code to fetch the credentials from the annotations
`AnnDeletionSecretRef*` if volumesnapshotclassname is not present.

Signed-off-by: Niraj Yadav <niryadav@redhat.com>