common-controller: fix empty `VolumeSnapshotRef.UID` in `VolumeSnapshotContent.Spec`
The function `createSnapshotsForGroupSnapshotContent` responsible for
creating `VolumeSnapshots` and binding them to the
`VolumeSnapshotContent` has a logic error where it reads an empty
`UID` for the `VolumeSnapshot` if it already exists.

This patch addresses the issue by fetching the existing
`VolumeSnapshot` if it already exists. That way the `UID` will
always be set by the API server. An error is returned to retry the
reconcile if an empty `UID` is found on this fetched object.

Signed-off-by: Niraj Yadav <niryadav@redhat.com>